+++ docs/change_log.rst
@@ -55,6 +55,7 @@ conversion (:trac:`14075`).
   * :class:`~stem.response.events.CircuitEvent` support for the new SOCKS_USERNAME and SOCKS_PASSWORD arguments (:trac:`14555`, :spec:`2975974`)
   * The 'strict' argument of :func:`~stem.exit_policy.ExitPolicy.can_exit_to` didn't behave as documented (:trac:`14314`)
   * Threads spawned for status change listeners were never joined on, potentially causing noise during interpreter shutdown
+  * Added support for specifying the authentication type and client names in :func:`~stem.control.Controller.create_hidden_service` (:trac:`14320`)
 
  * **Descriptors**